requirements.txt
Django==3.2.7
djangorestframework==3.12.4
psycopg2-binary
pillow==3.3.0